moving test classes back next to test method
